Move all components inside app to pages folder
- Move all redux files into redux folder
