--- src/components/app/index.js
@@ -1,2 +1,0 @@
-import App from './app'
-export default App
