+++ src/index.js
@@ -6,7 +6,7 @@ import 'bootstrap/dist/css/bootstrap.css';
 import App from './components/app';
 import * as serviceWorker from './serviceWorker';
 import { Provider } from 'react-redux'
-import store from './config/store'
+import store from './redux/store'
 
 ReactDOM.render(
   <React.StrictMode>
