+++ src/pages/movieDetails/movieDetails.css
