+++ src/pages/movieScreen/movie/index.js
