+++ src/pages/movieScreen/movie/movie.js
