+++ src/pages/movieScreen/searchBar/index.js
