+++ src/pages/movieScreen/searchBar/searchBar.css
