+++ src/pages/movieScreen/searchBar/searchBar.js
